move constants to avoid circular imports
